[IMP] l10n_uy_edi: Cancel doc if rejected by DGI
This commit aims to automatically set the status of a “Rejected”
document to “Cancelled” while preserving the sequence number assigned
by DGI (based on the CAE) to ensure all documents,
including rejected ones, remain in the system with their original sequence.
This is to maintain proper accounting records and provides visibility of
all rejected documents.

closes odoo/enterprise#80711

X-original-commit: 5af9024589d14f594085a162411586ca9950df61
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>