[FIX] l10n_it_reports: Modify the test to follow with the new reports
Description of the issue this commit addresses:

Since the splitting of the italian tax reports in this:
https://github.com/odoo/enterprise/commit/04392c34ddb585b95a26b6bf10ae4af352fefaea
the tests don't use the right report anymore.

Desired behavior after this commit is merged:

The tests use the right report.

Note on the forward port:

The FW is quite different from the original commit linked above. This is due to
a difference in the way integer rounding is added to reports. As from saas~17.2,
it can be done in community so this commit has a total different objective than
the original one.

Task-3479785

closes odoo/enterprise#77217

X-original-commit: 99074100ec5ed92ab75608a9a3b8891fde14734f
Related: odoo/odoo#193863
Signed-off-by: William André (wan) <wan@odoo.com>