[FIX] appointment: Hide attendee status indicator for non appointment
Status indicator was designed independently from the calendar indicator.
Overriding its behavior leads to the situation where we show appointment
status instead of the calendar indicator.

This commit attempts to keep the changes minimal so we are hiding
the status indicator for non-appointment events from the appointment template.

Reproduce
---
- Install appointment.
- Schedule a meeting in the calendar.
- Open the "Decline" link from the received confirmation email (as public_user).
- BUG: See "Appointment Reserved!" even tho we clicked "Decline"

Note: The bug doesn't occur when the appointment is not installed.

opw-4295448

closes odoo/enterprise#79321

X-original-commit: 505735a1d7279b8b71f63b1226e52601edf415b1
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>
Signed-off-by: Andrzej Pietrusiak (pian) <pian@odoo.com>