[IMP] *: gantt: fold off hours
This commit allows to fold gantt columns that are not "useful" on the
screen: this consists in consecutive columns that are filled with
unavailabilities on each row and that do not contain any pill as well.
Such columns therefore are folded into a single one when the user
activates the option and can be unfolded by clicking on any cell of
the folded column.

task-4350760

Part-of: odoo/enterprise#76951
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>