[FIX] test_l10n_be_hr_payroll_account: broken tour
During the upgrade to the "new" headless mode of Chrome (default for
v128+), an issue with the tours "hr_contract_salary_tour" and
"hr_contract_salary_tour_2" was detected.

In a nutshell, in those tours, during the signing of the contract, the
"street" field wasn't pre-filled and leaving it empty. The step-by-step
filling system (e.g. the "Next" button on the left) was confused and put
the date in the street field (sic). For some reason, the "old" Chrome
headless mode - while encountering the same issue - didn't complained
with it... But the "new" mode (which is actually closer to the actual
GUI behavior of Chrome) got stuck at those steps.

This commit fixes this issue by using the `private_street` field (which
contains the actual street name) instead of the `private_street2` field
(which is often empty).

Once done, a second issue was detected as the step filling the
computation of "58/4" didn't match (even wrongly) any empty field
anymore.

This commit simply removes those unused steps.

closes odoo/enterprise#78780

X-original-commit: 9f64f629a83c5f50a0e539b177da229c3753822c
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>