[FIX] sale_subscription: fix tour trigger selector
After the conversion of sale_sidebar to interaction, the element
"$spyWatcheElement" was removed. When generating the menu, this
element would have the id "1" and in the test, the optional
products section would have the id "3".

This commit updates a tour that used the id created while
generating the menu to replace id with the new one, computed
without take into account the "$spyWatchedElement".

closes odoo/enterprise#77844

Related: odoo/odoo#185998
Related: odoo/upgrade#7070
Signed-off-by: Géry Debongnie <ged@odoo.com>