[FIX] sale_subscription_timesheet: fix stale delivered qty on SOL
The issue:
The delivered qty is not reflected on the order line after creating
the first recurring invoice by the cron job.

Steps to reproduce:
1. Create a product with the following configuration:
	Product Type: Service
	Create on Order: Nothing
	Invoicing Policy: Based on Timesheets
	Product Category: Sellable and Subscription Product
2. Set up a recurring subscription plan for the product
3. Create a Sales Order (SO) for the product and confirm it
4. Add hours to a task linked to the SO
5. From Settings > Technical > Automation > Scheduled Actions,
   manually trigger the cron job "Sale Subscription: generate
   recurring invoices and payments" to update the `next_invoice_date`.

Observed behavior:
On the SO page, the smart button of the tasks reflects the hours added
correctly. However, the delivered quantity field in the SO line does not
update based on the added hours.

The reason:
The delivered quantity for timesheet based services is computed taking
into account the current period [1], however, `next_invoice_date` was not
added as a dependency to the compute method of `qty_delivered`, causing
this latter to have a stale value when `next_invoice_date` gets updated
when creating the first recurring invoice by `_create_recurring_invoice` [2].

The fix:
Added `next_invoice_date` as a dependency for the compute function of
`qty_deliverd`.

[1]: https://github.com/odoo/enterprise/blob/e990d101cce90311f3e2c5d9be4d480b329f9f2a/sale_subscription_timesheet/models/sale_order_line.py#L39
[2]: https://github.com/odoo/enterprise/blob/07f01f39cb46f1763637b50e628dfb613b0d2266/sale_subscription/models/sale_order.py#L1168

opw-4378914
opw-4426787

closes odoo/enterprise#77676

X-original-commit: 50a7e21721eacdf8eb8f66a4d9f40f8794bd49d3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>