[IMP] account_{batch_payment, online_payment}: KYC before payments
When a user create the first payment initiation with Ponto he need to do the KYC
of Ponto. The issue is that today we initiate the payment and then start the
process of KYC. Following that the first payment is always in error, because
it was initiated before the KYC.

In this commit, we create the payments only if the KYC is done.

task-4476856

closes odoo/enterprise#77348

X-original-commit: 7bfeb39f64bfecde2960c6f6678062ac515589cb
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>