[FIX] account_bank_statement_import_ofx,l10n_br_reports: allow non-unique FITID
To replicate the issue:

1. Install the brazilian localization (enterprise)
2. In the accounting app dashboard, click Import Statement
3. Upload an .OFX file with duplicate FIDID
4. A validation error appears and the upload is not done

Cause:
The bank statement import assumes the FIDID is a unique identifier
for the transactions in the .OFX file. A sql constraint is set in
place to ensure this uniqueness. However, some Brazilian banks may
generate files with transactions with the same FITID. An example
is when PIX transfers have the same FITID as the fee charged for
the PIX, which appears as another bank statement line with the
same FITID.

Fix:
The function _parse_bank_statement_file is altered to call a
function that fills the vals_line dictionary with the
transaction's values. This function is then overwritten in
l10n_br_reports to not include the FITID (unique_import_id) for
journals of Brazilian companies.

opw-4317681

X-original-commit: ef04c84a7c066d46fa44f9738c966e84f61b9112
Part-of: odoo/enterprise#77513
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>