[REF] mail, various: have email-like explicit external recipients
When posting a message on a thread, emails are split for recipients.
When sending emails to recipients, tune the Msg['To'] envelope of
outgoing emails so that they look like being sent to 'all external
people'. This can be done thanks to [1].

This does not change the SMTP To, hence the email is still sent only
to the specific recipient. But then the recipient can 'Reply All'
and reply to all customers + Odoo. Indeed Reply-All includes 'Reply-to'
in addition to original 'To'.

Update enterprise accordingly. See community PR for more details.

Task-4422660: [mail] Use external suggested recipients
Prepares Task-4273479: [mail] Email-like recipients

Part-of: odoo/enterprise#73273
Related: odoo/odoo#185240
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>