[IMP] web_enterprise: better call context menu
This commit fixes various issues with the context menu on the call
participant:

- Before this commit it wasn't possible to set the volume of a user to
0 due to the use of the `||` operator in the getVolume function. This
commit fixes the issue by using the correct `??` operator.
- Before this commit the volume slider in dark mode had the same
background as the popover menu, making it difficult to see. This
commit fixes the issue by making the slider darker than the
background.
- A volume icon was added to better indicate the slider purpose.
- A volume percentage tooltip was added for more accurate use.
- A chevron button was added to the participant card for  better
discoverability of the context menu.

closes odoo/enterprise#77352

X-original-commit: c66b2ff221929cc8da0a85f14f0f39a7f7042b14
Related: odoo/odoo#194151
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>