[FIX] documents: save all details panel changes
When editing a previewed document, no data was saved.

Reproduce:
1. Without selecting any document, open a preview
2. Rename the document or add a tag
3. Close the preview
4. Reload the page
5. See that changes were not saved

BTW, tags could not be created from the inspector, it's not clear
why.

Technical note: for previewed record, that are not "selected"
when we update them, we need to force saving immediately.
For records that are selected, we can leave the webclient
behavior for saving records (urgent save).
This can lead to having tracked fields being immediately
reported (seen in chatter) when editing previewed documents
but not the others. This is acceptable for simplicity.

Task-4465443

X-original-commit: be568230ed5a93820f5e8e8cdee7af899231286b
Part-of: odoo/enterprise#77767
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>