[FIX] documents: save all details panel changes
Several things addressed on the details panel.

A. When editing a previewed document, no data was saved.

Reproduce:
1. Without selecting any document, open a preview
2. Rename the document or add a tag
3. Close the preview
4. Reload the page
5. See that changes were not saved

B. When showing data for the current folder opened, the
contact info was missing (partner_id).

Reproduce:
1. Go inside any folder
2. Without selecting any documents, add a contact to the
  current folder via the details panel
3. Reload the page
4. The partner doesn't show up in the field (even though it is
  updated on the record).

C. Tags could not be created from the inspector, it's not clear
why.

Technical note regarding A: For previewed record, that are not
"selected" when we update them, we need to force saving
immediately. For records that are selected, we can leave
the webclient behavior for saving records (urgent save).
This can lead to having tracked fields being immediately
reported (seen in chatter) when editing previewed documents
but not the others. This is acceptable for simplicity.

Task-4465443

closes odoo/enterprise#77196

X-original-commit: 03d23c1f19241224175a8424d6fc06cc0692e294
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>