[IMP] crm_enterprise: generate leads from business cards images
This commit allows end users to upload business card images (or take a picture
on the fly) and generate leads from them.

The underlying logic relies on an API call to an AI service to read the image
and return a dictionary of information that we use to create lead records.

The AI service is IAP by default but can be configured to a custom openAI key.

Task-4210816

closes odoo/enterprise#83175

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>