[FIX] account_online_synchronization: Sudo when toggle the cron
The commit fixes the commit[[1]] that optimizes the cron activation
logic. Before this fix, only admin users could run the manual fetch,
it was wrong and we need to authorize all the users that have access
to the bank sync to active/deactivate this cron by using the manual fetch.

opw-4477366

[1]: https://github.com/odoo/enterprise/commit/bc3b9136d4d62fbc1d09267f73cac7ca29f1ba3a

closes odoo/enterprise#77226

X-original-commit: 75ae1612c5082e54198c74c75bf8427d0072ade5
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>