[REM] documents_spreadsheet: remove spreadsheet folder
We remove the spreadsheet folder as we make all spreadsheets lands into My
drive.

As we remove the spreadsheet folder, we move the demo data that were present in
that folder in the internal folder. We also change their access right to
access_internal='edit' do that they are visible and editable by internal user.

Task-4294237

Part-of: odoo/enterprise#74571
Related: odoo/odoo#191459
Related: odoo/upgrade#6820
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>