[IMP] sale_commission: display achievements in the current curency
Before this commit, the achievements were displayed in the currency of
the plan. As we want to be able to count achievements based on
moves/sales done in other company than the company of the plan, we
needed to convert the achieve amounts in a common currency. We decided
to use the currency of the current selected company.

Taskid: 4471762
X-original-commit: 374dda10260aaa271c9d6436438e50e5a8192c0d
Part-of: odoo/enterprise#79887
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>