[FIX] documents: keep translations when embedding action
When creating an embedded action from a server action,
only one language was copied.

Note that we are using `_get_stored_translations` instead of
`get_languages` to identify values to write as a simple
method to avoid storing too much unnecessary data as possible
(e.g., 100 languages installed but only `en_US` on the record).

Task-4486261

X-original-commit: 55326c2b68fa40682ab013a0edb27b1988ac7f91
Part-of: odoo/enterprise#77802
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>