[IMP] account_reports: comparison filter button more clear
Before this commit, the button label displayed as
'Comparison: <previous_period>' (where 'previous_period' was the
period preceding the compared period), regardless of the comparison
type selected. For example, if comparing October 2024 to previous
months, the label would appear as 'Comparison: Sep 2024'.
This caused confusion for users when comparing a period to multiple
previous ones, as only the last period was displayed in the label.

After this commit, the button label is now dynamic and reflects the
selected comparison:
- When comparing to previous periods:
    'Comparison: <periods_number> Previous Periods'
- When comparing to the previous period:
    'Comparison: Previous Period'
- When comparing to the same period in previous years:
    'Comparison: <periods_number> Previous Years'
- When comparing to the same period in the previous year:
    'Comparison: Previous Pear'
- When we don't compare to another period/year:
    'Comparison'
Here, 'periods_number' refers to the number of periods or years
being compared.

Task-id: 4267209
X-original-commit: 81cb58cb594d0681e22093d871d4011b7002407f
Part-of: odoo/enterprise#77506
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Jonathan Dauwe (jond) <jond@odoo.com>