[FIX] partner_comission: don't change UoM
During upgrades to 18.1, if the UoM of the product had been changed, it leads to
```
This product is already being used in posted Journal Entries.
If you want to change its Unit of Measure, please archive this product and create a new one.
```
No reason to revert user change during upgrade.

closes odoo/enterprise#77623

X-original-commit: d7d816efdff81c9c6eab1f11b04445f08c73e519
Signed-off-by: Thibault Francois <tfr@odoo.com>