[FIX] website_appointment: allows editing dynamic snippet link
Steps to reproduce:

- Install the appointments module.
- Go to the homepage.
- Enter edit mode.
- Drag and drop a "Appointments" block onto the page.
- Click on the "See all" link.
- Edit the link (e.g., change "/appointment" to "/contactus").
- Save the page.
- Click on the "See all" link.
- Bug: You are redirected to the "/appointment" page instead of the
"/contactus" page as expected.

The bug occurs since commit [1], where the "See all" link was added to
"Dynamic" snippets. Each time a "Dynamic" snippet is rendered during a
page load, the href of the "See all" link is reset to the main module
page. While this behavior is useful when the snippet is first rendered
after being dropped onto the page, it is problematic afterward. Once the
user has edited the link with a different href, it should not be reset
on every page load.

[1]: bbcff74

opw-4418048

closes odoo/enterprise#77284

Related: odoo/odoo#193983
Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>