[IMP] {website_}knowledge: move html_field migration to html_editor
Removing the Excalidraw command means that the html migration system must be
moved to `html_editor`.
This commit thus removes the sytem from Knowledge and adapts the manifests to
take into account the move done, and updates a test to account for the new html
editor version (1.1).

task-4489076

closes odoo/enterprise#80003

X-original-commit: ad179ad4b000a5bf392444b6caced86e7e130c25
Related: odoo/odoo#198942
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>