[FIX] account_auto_transfer: Fix transfer UI
The transfer form view contains two list views that were poorly
displayed. The two were shrunk to half the available space.

This is due to the fact that a group is by default two divs in a row
each expecting one element, usually the label and the field value.
However, only the value is to be displayed in our case. Hence, why the
list only take up half of the space made avalailable by the group.

To solve this, we can simply force the number of columns taken by the
list to 2.

task-4481314

closes odoo/enterprise#77388

X-original-commit: 50aba169a0a8e489de1cc2204de850dad9a1f01b
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>