[FIX] stock_barcode: allow creation of lotless quant with quantity
Steps to reproduce:

- Create a storable product tracked by lot/serial
- Click on "On Hand"
- Create a new line with 10 "On Hand Quantity" WITHOUT lot/serial
- Go back to the product: The "On Hand" quantity is still at 0.
- Click back to the "On Hand": the quantity of the line is back to 0.

Cause of the issue:

When a new line is created and the "On Hand Quantity" is edited from the
`view_stock_quant_tree` view, we actually performs a create of the
stock.quant followed by a write on the `inventory_quantity_auto_apply`.
At this point the 'quantity' field of the new quant is not set and
the `action_apply_inventory` of the `inventory_quantity_auto_apply`
is expected to create and validate a move line in order to update our
quant quantities here:
https://github.com/odoo/odoo/blob/e041e890b91e5bc515e1827683f073781788f253/addons/stock/models/stock_quant.py#L230-L235
However, this steps will never be performed as we return the call before
the `_apply_invetory` because our quant is set without lots:
https://github.com/odoo/odoo/blob/e041e890b91e5bc515e1827683f073781788f253/addons/stock/models/stock_quant.py#L447-L449
https://github.com/odoo/odoo/blob/e041e890b91e5bc515e1827683f073781788f253/addons/stock/models/stock_quant.py#L465-L477

Master Fix:

We remove the wizzard.

opw-4428050

closes odoo/enterprise#77344

Related: odoo/odoo#194079
Related: odoo/upgrade#7088
Signed-off-by: William Henrotin (whe) <whe@odoo.com>