[PERF] spreadsheet: load only one bundle
See community PR.

Currently, when we load a spreadsheet client action, we load two asset
bundles:  'web.chartjs_lib' and 'spreadsheet.o_spreadsheet'

It's 2 network requests per bundle (or only 1 once the asset js file is cached
by the browser).

We can't load both in parallel because 'spreadsheet.o_spreadsheet' needs
ChartJs to be loaded to register some plugins.

So currently, before loading 'spreadsheet.o_spreadsheet', we wait for 2
sequential network calls, which can take a significant amount of time
(hundreds of ms) depending on the network latency.

With this commit, the bundle 'web.chartjs_lib' is included in the main
'spreadsheet.o_spreadsheet'. The amount of javascript loaded is the same
but it's loaded in half as many requests.

closes odoo/enterprise#77474

Task: 4497439
X-original-commit: 12f1dc6d772c164c9fce42ae61092b9087c9fa8e
Related: odoo/odoo#194479
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>