[IMP] sale_subscription: don't create the invoice in portal controller
Before this commit, the /my/subscriptions/<int:order_id>/transaction
route would create a draft invoice when it was called to anticipate a
payment. post processing of payment.transaction is able to create the
invoice and to associate it with the correct sale.order.

In order to simplify the code by limiting the part of the code where
subscription invoices can be created, we decided to let the
_post_process do the work.

closes odoo/enterprise#78865

Taskid: 4368037
X-original-commit: e628b30adea2e19404519515fd6da388a5877778
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>