[FIX] account_asset: allow to create asset models
Steps to reproduce:
- Install Accounting
- Go to "Accounting / Configuration / Management / Asset Models"
- Create an asset model

Issue:
Upon save, the following UserError is raised:
"You cannot add or remove bills when the asset is already running
or closed."

Cause:
A check has been added to prevent adding bills to a running asset.
The check is excluding draft assets but not asset models that are
assets in "model" state.

Issue:
Exclude asset models from the check.

opw-4479698

closes odoo/enterprise#77662

X-original-commit: fe5b4fca2f2d778838a67565b20d2d513161ca8b
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>