[IMP] web_gantt: simplify expand/collapse buttons
We make only one of the buttons "Expand rows"/"Collapse rows" appear and use
different icons. Now the button "Collapse rows" only appears when no row is
folded. For this to work correctly, we have to clear more often the set
closedRows. Roughly speaking, we clear that set in such a way that an automatic
expand happens when it would happen in similar situations in the list view.

Task ID: 4500469

closes odoo/enterprise#77479

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>