[REF] website_knowledge: make public view more reactive
Purpose:
-------
The current public view of knowledge has several flaws:
- clicking on an article in the sidebar redirects to that article, which
results in a complete page reload and a flicker.
- the sidebar's html is generated in backend and inserted in the dom,
which makes it non reactive and slow to load.
- searching for an article also generates a list of articles in backend
which is inserted in the dom, and only matches article titles.

To alleviate these issues, this commit introduces:
- a minimalist `Article` class used as the article record in the public
view. Changing article now fetches the article's data through a rpc
instead of reload the public view template with the new article, avoiding
the flicker and making it faster.
- a component `knowledge_public_sidebar` that loads the data of the articles
and render the sidebar in frontend.
- a articleSearchDialog similar to the backend's command palette, which
allows to find articles of the current subsite based on their title or
their content.

Some simplifications and feature changes have been done:
- the unfolded article ids are not stored in the local storage anymore:
main purpose was to keep the articles unfolded when clicking on an
article in the sidebar, since the page was reloaded. But now since the
sidebar isn't reloaded on article change, it's not necessary anymore.
- the sidebar can be toggled in dekstop mode too.
- the "load more" feature has been removed, as it was introduced for
preventing the view from crashing. The new component approach prevents
these crashes from happening.

Task-4439155

closes odoo/enterprise#77597

Related: odoo/odoo#197632
Related: odoo/upgrade#7130
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>